requirements.txt
numpy                # array computations
torch                # tensor computations and deep learning
Pillow               # image processing
matplotlib           # data visualization